refactor: moved ApplicationItem component to a separate file
